+++ CHANGELOG.md
@@ -17,6 +17,13 @@ and this project adheres to
 
 ### Added
 
+### Changed
+
+### Fixed
+
+- Fix provioning API calls workflow limiter without the project ID
+  [#2057](https://github.com/OpenFn/lightning/issues/2057)
+
 ## [v2.4.4] - 2024-05-03
 
 ### Added